refactor: 모바일페이지 컴포넌트 분리 (#477)
* feat: 데이터 추가 mock api 연결

* refactor: 인증 필요 없는 기능 client 변경

* feat: mse 동영상, 리뷰 기능 구현 (#472)

* refactor: api 훅 함수로 분리, 적용 (#472)

* refactor: 쓰지 않는 파일 삭제 (#472)

* fix: console log 삭제 (#472)

* Squashed commit of the following:

commit 6e0068735b155b95ec45de7fd78eb241993dc513
Author: Jeremy <102432453+shackstack@users.noreply.github.com>
Date:   Fri Sep 15 17:18:33 2023 +0900

    bug: 옵션관련 ux 개선 (#474)

    * fix: 검색창 옵션 수정 (#473)

    * fix: 지도 줌 limit 설정 (#473)

    * fix: 모바일 환경에서 거리순으로 설정 (#473)

    * design: 오버레이 마커 디자인 개선 (#473)

* fix: 오류 수정 (#476)

* refactor: useScrollBlock 보강 (#476)

* refactor:  모바일 페이지 컴포넌트 분리 (#476)

* refactor:  라우터 재설정 (#476)

* fix: 잘못된 import 수정 (#476)

* refactor: router 및 outlet 새로 설정 (#476)

---------
