+++ frontend/src/components/BottomNavBar/index.tsx
@@ -1,0 +1,3 @@
+import BottomNavBar from './BottomNavBar';
+
+export default BottomNavBar;
