+++ frontend/src/hooks/useScrollBlock.ts
@@ -1,13 +1,19 @@
 import { useEffect } from 'react';
 
-const useScrollBlock = (refs: React.MutableRefObject<Element>[]) => {
+type Ref = React.MutableRefObject<Element>;
+
+const useScrollBlock = (refs: Ref[] | Ref) => {
   const blockScroll = (ref: React.MutableRefObject<Element>) => {
     ref.current.addEventListener('touchmove', e => e.preventDefault(), { passive: false });
   };
 
   useEffect(() => {
-    refs.forEach(ref => blockScroll(ref));
-  }, []);
+    if (Array.isArray(refs)) {
+      refs.forEach(ref => blockScroll(ref));
+    } else {
+      blockScroll(refs);
+    }
+  }, [refs]);
 };
 
 export default useScrollBlock;
