+++ frontend/src/pages/WishListPage.tsx
@@ -1,10 +1,5 @@
-import { Link } from 'react-router-dom';
-
 import styled, { css } from 'styled-components';
-import Logo from '~/assets/icons/logo-icon.svg';
 
-import Footer from '~/components/@common/Footer';
-import Header from '~/components/@common/Header';
 import RestaurantWishList from '~/components/RestaurantWishList';
 import LoginErrorHandleComponent from '~/components/@common/LoginErrorHandleComponent';
 
@@ -17,24 +12,10 @@ function WishListPage() {
   return (
     <LoginErrorHandleComponent>
       <StyledWishListPageWrapper>
-        <div>
-          {isMobile ? (
-            <StyledMobileHeader>
-              <Link aria-label="셀럽잇 홈페이지" role="button" to="/">
-                <Logo width={32} />
-              </Link>
-              <h5>celuveat</h5>
-              <div />
-            </StyledMobileHeader>
-          ) : (
-            <Header />
-          )}
-          <StyledMobileLayout>
-            <StyledTitle isMobile={isMobile}>위시리스트</StyledTitle>
-            <RestaurantWishList />
-          </StyledMobileLayout>
-        </div>
-        <Footer />
+        <StyledMobileLayout>
+          <StyledTitle isMobile={isMobile}>위시리스트</StyledTitle>
+          <RestaurantWishList />
+        </StyledMobileLayout>
       </StyledWishListPageWrapper>
     </LoginErrorHandleComponent>
   );
@@ -71,15 +52,4 @@ const StyledTitle = styled.h3<{ isMobile: boolean }>`
     css`
       font-size: ${FONT_SIZE.lg};
     `}
-`;
-
-const StyledMobileHeader = styled.header`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  width: 100%;
-  height: 44px;
-
-  padding: 0.2rem 0.8rem;
 `;
